# chore: Improve containsLower performance using quick rejection